backend/requirements.txt
opencv-python
opencv-contrib-python
ultralytics
flask
deep-sort-realtime
aiortc
av
aiohttp
sqlalchemy
alembic
psycopg2-binary
python-dotenv
